requirements.txt
pytest
Faker
selenium
requests
allure-pytest